POC eth-accounts module refactoring
